apps/fact-checker/pyproject.toml
[project]
authors = [
  {name = "Bharath Lakshman Kumar", email = "bharathxxd@gmail.com"},
]
dependencies = [
  "langchain (>=0.3.25,<0.4.0)",
  "langchain-openai (>=0.3.16,<0.4.0)",
  "nltk (>=3.9.1,<4.0.0)",
  "langgraph (>=0.4.3,<0.5.0)",
  "python-dotenv (>=1.1.0,<2.0.0)",
  "langgraph-sdk (>=0.1.69,<0.2.0)",
  "langchain-tavily (>=0.1.6,<0.2.0)",
]
description = "A fact-checking agent that extracts claims from a given text and validates them as complete, declarative sentences."
name = "fact-checker"
readme = "README.md"
requires-python = ">=3.11,<4.0"
version = "0.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
create-run = "scripts.create_run:main"
